Make the FogRemover stateless
